Query Builder Interaction and Validation (#3520)
* add logic to build expression trees

* Get started with keyboard interaction tests

* push pair state

* Pairing v2


* add rendermodelindex concept to the tests

* Test focusNext

* add rainbow colors to expressions

* add controls to the story book so we can easily create complex use cases

* Complete tests for left/right keyboard interaction

* add clear button

* use inline to get a proper wrap of tags. also only remove style from the last exapression item if it is a closing tag

* add margin to space to get a proper wrap

* add margin to space to get a proper wrap

* remove bg

* Support addition of tags when focussing spaces, add a blinking cursor and support jump to beginning/end

* Also happy terminal users happy

* Pairing

* enhance tag

* add operator overlay

* Support positioning of cursor when clicking behind the last element in a line

* Support selection of operators with full keyboard interaction

* Full keyboard support for conjunction selector

* add basic validation to the render model

* there is a letter at the end now

* get operator change working

* validation information is only available on invalid elements

* add doku about ending letter

* add coloring to tags to see if they are valid or not

* Support keyboard interaction for brackets

* always create a copy of elements when transforming form to render model

* always show expression colors

* lighten the expression colors

* Get started with form model and tag validation

+++ packages/in-new-components/QueryBuilder/components/Spacing.js
@@ -1,40 +1,60 @@
 import React from 'react';
 
 import { LETTER, WORD } from 'in-new-components/QueryBuilder/transformation/renderModel';
+import { isDefaultInteractionTrigger } from 'in-new-components/interactiveCustomElement';
 import { stopPropagationAndPreventDefault } from 'in-services/util/function';
 import { evaluateClassNames } from 'in-services/util/classnames';
 import keyCodes from 'in-components/keyCodes';
+import SvgIcon from 'in-components/SvgIcon';
 
 import locals from './Spacing.mless';
 
-export default function Spacing({ renderModelIndex, size, onRemove, leftFormModelIndex, rightFormModelIndex }) {
+export default function Spacing({ renderModelIndex, size, valid, onRemove, leftFormModelIndex, rightFormModelIndex }) {
   return (
     <div
       className={evaluateClassNames({
         [locals.letter]: size === LETTER.size,
-        [locals.word]: size === WORD.size
+        [locals.word]: size === WORD.size,
+        [locals.invalid]: valid === false
       })}
       tabIndex={0}
       data-render-model-index={renderModelIndex}
       data-query-builder-element="true"
+      onKeyDown={onKeyDown}
       onKeyUp={onKeyUp}
     >
+      <div className={locals.addIndicator} onClick={onOpenAddDialog}>
+        <SvgIcon type="lib_openclose_add" className={locals.addIndicatorIcon} size="xxs" />
+      </div>
       &nbsp;
     </div>
   );
 
   function onKeyUp(e) {
-    if (!onRemove) {
-      return;
-    }
-
-    if (e.keyCode === keyCodes.backspace) {
+    if (isDefaultInteractionTrigger(e)) {
+      stopPropagationAndPreventDefault(e);
+      onOpenAddDialog();
+    } else if (onRemove && e.keyCode === keyCodes.backspace) {
       stopPropagationAndPreventDefault(e);
       // Deleting a single element also deletes the whitespace element.
       onRemove(leftFormModelIndex, renderModelIndex - 2);
-    } else if (e.keyCode === keyCodes.delete) {
+    } else if (onRemove && e.keyCode === keyCodes.delete) {
       stopPropagationAndPreventDefault(e);
       onRemove(rightFormModelIndex, renderModelIndex);
     }
+  }
+
+  function onOpenAddDialog() {
+    // eslint-disable-next-line
+    console.log('TODO show add tag dialog');
+  }
+}
+
+function onKeyDown(e) {
+  if (isDefaultInteractionTrigger(e)) {
+    // Need to stop this already in onKeyDown as otherwise the browser
+    // may start to scroll when using the spacebar to trigger the default
+    // interaction.
+    stopPropagationAndPreventDefault(e);
   }
 }
